Move pulumi components to their folder
