+++ infrastructure/components/static-site.ts
